+++ cli/cli/kurtosis_gateway/server/api_container_gateway/api_container_gateway_service_server.go
@@ -341,6 +341,7 @@ func (service *ApiContainerGatewayServiceServer) startRunningConnectionForKurtos
 			TransportProtocol:        privateApiPort.GetTransportProtocol(),
 			MaybeApplicationProtocol: privateApiPort.GetMaybeApplicationProtocol(),
 			MaybeWaitTimeout:         privateApiPort.GetMaybeWaitTimeout(),
+			Locked:                   nil,
 		}
 	}