+++ enclave-manager/api/typescript/src/kurtosis_enclave_manager_api_connect.ts
@@ -3,10 +3,11 @@
 /* eslint-disable */
 // @ts-nocheck
 
-import { CreateRepositoryWebhookRequest, DownloadFilesArtifactRequest, GetListFilesArtifactNamesAndUuidsRequest, GetServicesRequest, GetStarlarkRunRequest, HealthCheckRequest, HealthCheckResponse, InspectFilesArtifactContentsRequest, RunStarlarkPackageRequest, RunStarlarkScriptRequest, StarlarkPackagePlanYamlArgs, StarlarkScriptPlanYamlArgs } from "./kurtosis_enclave_manager_api_pb.js";
+import { CreateRepositoryWebhookRequest, DownloadFilesArtifactRequest, GetListFilesArtifactNamesAndUuidsRequest, GetServicesRequest, GetStarlarkRunRequest, HealthCheckRequest, HealthCheckResponse, InspectFilesArtifactContentsRequest, LockUnlockPortRequest, RunStarlarkPackageRequest, RunStarlarkScriptRequest, StarlarkPackagePlanYamlArgs, StarlarkScriptPlanYamlArgs } from "./kurtosis_enclave_manager_api_pb.js";
 import { Empty, MethodKind } from "@bufbuild/protobuf";
 import { CreateEnclaveArgs, CreateEnclaveResponse, DestroyEnclaveArgs, GetEnclavesResponse, GetServiceLogsArgs, GetServiceLogsResponse } from "./engine_service_pb.js";
 import { GetServicesResponse, GetStarlarkRunResponse, InspectFilesArtifactContentsResponse, ListFilesArtifactNamesAndUuidsResponse, PlanYaml, StarlarkRunResponseLine, StreamedDataChunk } from "./api_container_service_pb.js";
+import { GetCloudInstanceConfigResponse } from "./kurtosis_backend_server_api_pb.js";
 
 /**
  * @generated from service kurtosis_enclave_manager.KurtosisEnclaveManagerServer
@@ -146,6 +147,33 @@ export const KurtosisEnclaveManagerServer = {
     createRepositoryWebhook: {
       name: "CreateRepositoryWebhook",
       I: CreateRepositoryWebhookRequest,
+      O: Empty,
+      kind: MethodKind.Unary,
+    },
+    /**
+     * @generated from rpc kurtosis_enclave_manager.KurtosisEnclaveManagerServer.GetCloudInstanceConfig
+     */
+    getCloudInstanceConfig: {
+      name: "GetCloudInstanceConfig",
+      I: Empty,
+      O: GetCloudInstanceConfigResponse,
+      kind: MethodKind.Unary,
+    },
+    /**
+     * @generated from rpc kurtosis_enclave_manager.KurtosisEnclaveManagerServer.LockPort
+     */
+    lockPort: {
+      name: "LockPort",
+      I: LockUnlockPortRequest,
+      O: Empty,
+      kind: MethodKind.Unary,
+    },
+    /**
+     * @generated from rpc kurtosis_enclave_manager.KurtosisEnclaveManagerServer.UnlockPort
+     */
+    unlockPort: {
+      name: "UnlockPort",
+      I: LockUnlockPortRequest,
       O: Empty,
       kind: MethodKind.Unary,
     },