+++ internal_testsuites/golang/test_helpers/test_helpers.go
@@ -127,18 +127,21 @@ var fileServerPortSpec = &kurtosis_core_rpc_api_bindings.Port{
 	TransportProtocol:        kurtosis_core_rpc_api_bindings.Port_TCP,
 	MaybeApplicationProtocol: emptyApplicationProtocol,
 	MaybeWaitTimeout:         defaultWaitTimeoutForTest,
+	Locked:                   nil,
 }
 var datastorePortSpec = &kurtosis_core_rpc_api_bindings.Port{
 	Number:                   uint32(datastore_rpc_api_consts.ListenPort),
 	TransportProtocol:        kurtosis_core_rpc_api_bindings.Port_TCP,
 	MaybeApplicationProtocol: emptyApplicationProtocol,
 	MaybeWaitTimeout:         defaultWaitTimeoutForTest,
+	Locked:                   nil,
 }
 var apiPortSpec = &kurtosis_core_rpc_api_bindings.Port{
 	Number:                   uint32(example_api_server_rpc_api_consts.ListenPort),
 	TransportProtocol:        kurtosis_core_rpc_api_bindings.Port_TCP,
 	MaybeApplicationProtocol: emptyApplicationProtocol,
 	MaybeWaitTimeout:         defaultWaitTimeoutForTest,
+	Locked:                   nil,
 }
 
 type GrpcAvailabilityChecker interface {